+++ iter_model/sync_iter.py
@@ -369,6 +369,18 @@ class SyncIter(Generic[T]):
             count += 1
         return count
 
+    @sync_iter
+    def batches(self, batch_size: int) -> 'SyncIter[tuple[T, ...]]':
+        """Create iterator of tuples whose length = batch_size"""
+        while True:
+            try:
+                item = self.next()
+            except StopIteration:
+                break
+            it = self.append_left(item)
+            batch = it.take(batch_size).to_tuple()
+            yield batch
+
     def __len__(self) -> int:
         return self.get_len()